% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irea-function.R
\name{GeneSetEnrichmentScore}
\alias{GeneSetEnrichmentScore}
\title{IREA analysis for gene set input - rank sum test method}
\usage{
GeneSetEnrichmentScore(degs, input_celltype, species = "mouse")
}
\arguments{
\item{degs}{A list of differentially expresssed genes that user would like to investigate}

\item{input_celltype}{Choose from one of the listed cell types that most resembles the input}
}
\description{
\code{GeneSetEnrichmentScore} Perform statistical test (Wilcoxon rank sum test) for cytokine response enrichment when the user input is a list of genes
}
