% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irea-visualization.R
\name{IreaCompassPlot}
\alias{IreaCompassPlot}
\title{Compass plot}
\usage{
IreaCompassPlot(df_irea, color_by = c("pval", "ES"), plot_receptor = FALSE)
}
\arguments{
\item{df_irea}{IREA results data frame returned from one of the "GeneSetEnrichmentScore",
"GeneSetEnrichmentHyperTest", or "GetEnrichmentScoreProjection" methods}

\item{color_by}{Color the IREA compass plot by p-value or effect size; default is p-value
(TODO: add support for ES)}
}
\description{
\code{IreaCompassPlot} Visualize IREA results on a compass plot
}
