% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irea-visualization.R
\name{IreaRadarPlot}
\alias{IreaRadarPlot}
\title{Radar plot}
\usage{
IreaRadarPlot(
  df_irea,
  color_by = "ES",
  analysis_type = "Polarization",
  input_celltype = "Macrophage",
  fill_color = "default"
)
}
\arguments{
\item{df_irea}{IREA results data frame returned from one of the "GeneSetEnrichmentScore",
"GeneSetEnrichmentHyperTest", or "GetEnrichmentScoreProjection" methods}

\item{color_by}{Color the IREA compass plot by p-value or effect size; default is p-value
(TODO: add support for ES)}

\item{analysis_type}{Visualize enrichment of immune cell polarization states or individual cytokines}

\item{fill_color}{Color to be filled on the radar plot}

\item{celltype}{Cell type analyzed}
}
\description{
\code{IreaRadarPlot} Visualize IREA results on a radar plot. Default visualization for immune cell polarization.
}
