% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irea-function.R
\name{PolarizationGeneSetEnrichmentHyperTest}
\alias{PolarizationGeneSetEnrichmentHyperTest}
\title{IREA polarization analysis for gene set input - Fisher's test method}
\usage{
PolarizationGeneSetEnrichmentHyperTest(degs, input_celltype, species = "mouse")
}
\arguments{
\item{degs}{A list of differentially expresssed genes that user would like to investigate}

\item{input_celltype}{Choose from one of the listed cell types that most resemble the input}
}
\description{
\code{PolarizationGeneSetEnrichmentHyperTest} Compute enrichment score using the pathway overrepresentation test,
the most commonly used method for pathway analysis which uses the hypergeometric test
}
