% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irea-function.R
\name{PolarizationGeneSetEnrichmentScore}
\alias{PolarizationGeneSetEnrichmentScore}
\title{IREA Polarization analysis for gene set input - rank sum test method}
\usage{
PolarizationGeneSetEnrichmentScore(degs, input_celltype, species = "mouse")
}
\arguments{
\item{degs}{A list of differentially expresssed genes that user would like to investigate}

\item{input_celltype}{Choose from one of the listed cell types that most resembles the input}
}
\description{
\code{PolarizationGeneSetEnrichmentScore} Perform statistical test (Wilcoxon rank sum test) for polarization enrichment when the user input is a list of genes
}
